gateway/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
PyJWT[crypto]>=2.8.0
supabase>=2.12.0
litellm>=1.60.0,<2.0.0
pydantic-settings>=2.7.0
sse-starlette>=2.2.0
svix>=1.21.0
websockets>=14.0

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.24.0
